Rework the header: bigger logo, nav and tagline sharing a row above the ticker
Logo alone occupies the top of the hero photo, sized up so its own
height pushes the row below it down to where the tagline sits - nav and
tagline now share that row, bottom-aligned, right above the ticker,
rather than three separate stacked rows leaving dead photo space
between them.

"Last updated" moved out of the hero entirely, onto the same line as
the "Latest headlines" caption instead - contextually closer to the
content it describes, and that caption is now a real gold heading
rather than a small muted label. The tagline itself went through a few
rounds of "bigger", then "lighter, that's too bold" - it's small and
medium-weight now, deliberately not competing with the logo for
attention.

Schedule widget: weekdays spelled out (formatGameDateTime in dates.js)
and "Week N" no longer shouts in caps.

+++ src/lib/dates.js
@@ -74,7 +74,7 @@ export function parseGameTime(value) {
 export function formatGameDateTime(iso) {
   const date = iso ? new Date(iso) : null;
   if (!date || Number.isNaN(date.getTime())) return '';
-  const day = date.toLocaleDateString('en-US', { timeZone: TZ, weekday: 'short', month: 'short', day: 'numeric' });
+  const day = date.toLocaleDateString('en-US', { timeZone: TZ, weekday: 'long', month: 'short', day: 'numeric' });
   const time = date.toLocaleTimeString('en-US', { timeZone: TZ, hour: 'numeric', minute: '2-digit' });
   return `${day} · ${time}`;
 }
